refatorando comadas da rota /UserPages
