+++ controllers/MenuItems.js
@@ -8,14 +8,6 @@ const getMenuItems = async (req, res) => {
   return res.status(200).json(itemsMenuWithPages);
 };
 
-const getPagesAllowed = async (req, res) => {
-  const userInfo = req.userAuthenticated;
-  const pagesAllowed = await MenuItems.getPagesAllowedByPerfil(userInfo.idPerfil);
-
-  return res.status(200).json(pagesAllowed);
-};
-
 module.exports = {
   getMenuItems,
-  getPagesAllowed,
 };
