+++ service/MenuItems.js
@@ -1,8 +1,6 @@
-const { QueryTypes } = require('@sequelize/core');
 const { 
   page, menu_item: MenuItem, 
   acess_permission: AcessPermission, 
-  sequelize, 
 } = require('../models');
 
 const { MenuItems: MenuItemsSchema } = require('../schemas');
@@ -32,24 +30,8 @@ const getAllItemMenuWithPagesAllowed = async (idPerfil) => {
   }
 };
 
-const getPagesAllowedByPerfil = async (idPerfil) => {
-  try {
-    const pages = await sequelize
-      .query(`SELECT id, name, route FROM tracker_manager.acess_permissions
-    INNER JOIN tracker_manager.pages as pg on id_page = pg.id
-    WHERE id_perfil = :idPerfil;`, {
-      replacements: { idPerfil },
-      type: QueryTypes.SELECT,
-    });
-    return pages;
-  } catch (error) {
-    console.log(error.message);
-  }
-};
-
 module.exports = {
   getAllItemMenuWithPagesAllowed,
   getAllItemsMenuWithPages,
   getAllPagesPermissionsByPerfil,
-  getPagesAllowedByPerfil,
 };
