+++ service/index.js
@@ -1,7 +1,9 @@
 const User = require('./User');
 const MenuItems = require('./MenuItems');
+const Pages = require('./Pages');
 
 module.exports = {
   User,
   MenuItems,
+  Pages,
 };
